+++ demo/button.gd
@@ -37,3 +37,7 @@ func _on_reset_button_up():
 func _on_loading_scene_button_up():
 	SceneManager.set_recorded_scene(scene)
 	SceneManager.change_scene("loading", fade_out_options, fade_in_options, general_options)
+
+func _on_loading_scene_initialization_button_up():
+	SceneManager.set_recorded_scene(scene)
+	SceneManager.change_scene("loading_with_initialization", fade_out_options, fade_in_options, general_options)